record has bug.noise get out but i don't know why!

+++ app/src/main/java/com/yibogame/superrecorder/interfaces/IOnRecordingListener.java
@@ -5,5 +5,5 @@ package com.yibogame.superrecorder.interfaces;
  */
 
 public interface IOnRecordingListener {
-    void onDataReceived(short[] mPCMBuffer, int readSize, double volume);
+    void onDataReceived(byte[] mPCMBuffer, int readSize, double volume);
 }
